FeedFetcher checks if item is already present
Although there is already a validation on the item ensure duplicate
items are not added from a feed, performance could be slow
as it would build the item, try to save and rollback.

Now it does a quick check in the loop to see if present and goes to the
next entry if that's the case.