Lets user set last_fetched_at on adding feed
Previously, when a new feed was added it would attempt to fetch
all entries available for an item. Depending on the feed this could be
just a handful of entries (e.g. Daring Fireball: 48) or it could
be hundreds (Six Colors: 400+).

Now when adding a feed, a user can choose how far back to initially
get items (24 hours - 2 weeks). This is achieved by passing in the
value to set on the feed and then adding a break clause which stops
looping through entries when it comes across an entry that was added
before the last_fetched_at time.

A positive side effect is that this will greatly reduce the response
time for completly a feed refresh as it will break off when it reaches
feed items it will have already imported on the last feed import.