Adds column of feed items
Instead of using a continuous feed that is scrolled through
items will are listed in a sidebar. This lays the groundwork for
each an item being clicked on and routed to the show page for that
resource.

It removes the infinite scroll for now and currently just shows
the first item in the feed.