Refactors generation of item paths
The prior commit introduced a bug where feed item links would
always link to the nested /feed route. This is because of the
front-end code that checks for the presence of a feed object. This
feed object is now always present so was no longer suitible for
selecting which sort of url to use. Additionally, it was icky that
this was done in the front-end.

Now leveraging Item serializers with params to generate the correct
item_path or feed_item_path. Removes the front-end complexity, even
if it does make the controller a bit more bloated.