Redirects to next item when item read/unread
Adds a Navigator class that is responsible for figuring out what the
next item is, and generates the path to that item. This is complicated
because there are one of four contexts for an item:

- all items
- all items in a feed
- all unread items
- all unread items in a feed

This Navigator class also sets the groundwork for getting previous,
first, last etc items.

Additionally, it adds a brakeman.ignore file that covers a
false-postive from brakeman. The navigator does not directly use user
params to set a redirect_url so is safe.