Adds refreshing a single feed with toast success
Due to the architecture of the general Dropdown component expecting
links, it wasn't practical to use that component, so for now pulled
out a specific component for the feed management dropdown. Will look
to refactor the Dropdown component to be more flexible in the future.

Does not remove flash messages – this will be done globally in a
seperate commit.