Reloads page 5 seconds after refres all feeds
This is a hack. Currently no callback to know when all the feeds
have refreshed. So basically, give it five second and do an inertia
page reload. This will repopulate the items list but it will
preserve the state / scroll position so it will not interrupt the
reading experience. Main issue is that 5 second is a random number.