Redesigns and refactors Dropdown
Prior implementation of Dropdown had a series of flaws
- Item implementation was coupled to an array that was passed in as
props. This caused issues when some items may be a Button or a Link.
– Didn't handle overflows
– Looked bad
– Used default outline styles when active

This PR fixes this with the following:
– Items are simply passed in as children. The downside is that there
is more boilterplate code for each item. This can be abstracted away
over time.
– Now leverages a Portal with Floating UI for placement to solve
overflow issue
– Overall improved design.

Also reduces the size of the header icons / links.