Adds account unread items via counter cache
This required switching to a direct association between account
and item for the counter cache to work (i.e. not relying on the
implicit through on the feed model). In turn there is a lot of
refactoring of the test suite, but I think this will be more flexible
anyway in the long run.