Adds simple item pagination
Main complexity is figuring out what page number an item is on and
getting that page's items. Not sure if the trade-off is right. Maybe
it would be better if it just went to the first paginated item if you
go to a an item direct... not even sure how often it matters.

For now the goal is to just stop grabbing every item and keep
response times and memory usage lower. Need to think a bit more
about how to get infinite scrolling in both directions sustainably.