Refactors feed mgmt page to derive state from url
Old solution was setting state, then doing request based on that state.
This solution does request, and derives local component state based on
the URL instead of having state effectively in two places.

This means, removing additional local state and a useEffect hook that
was running on render (so doing an extra request – useRef would have
solved this but been a mess).